Initial custom variables and page templating implementation.
This commit adds the ability to have arbitrary custom values on post
templates, that do not override the standard values already present in
Gempost.

It also adds the ability to have templated pages, separate from static
content and Gemlog posts. Initial implementation is to have one
template for all pages, but the intention is to allow templating for
different subdirectories or even individual pages.